feat: Refactor mossy-device to mossy-key-sync, updating package names and configurations
