+++ mossy-key-sync/src/main/kotlin/pl/dayfit/mossykeysync/configuration/RedisConfiguration.kt
@@ -1,8 +1,8 @@
-package pl.dayfit.mossydevice.configuration
+package pl.dayfit.mossykeysync.configuration
 
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.redis.repository.configuration.EnableRedisRepositories
 
 @Configuration
-@EnableRedisRepositories(basePackages = ["pl.dayfit.mossydevice.repository.redis"])
+@EnableRedisRepositories(basePackages = ["pl.dayfit.mossykeysync.repository.redis"])
 class RedisConfiguration
