+++ mossy-key-sync/src/main/kotlin/pl/dayfit/mossykeysync/configuration/SecurityConfiguration.kt
@@ -1,4 +1,4 @@
-package pl.dayfit.mossydevice.configuration
+package pl.dayfit.mossykeysync.configuration
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
