+++ mossy-key-sync/src/main/kotlin/pl/dayfit/mossykeysync/configuration/SockJsWebSocketConfiguration.kt
@@ -1,4 +1,4 @@
-package pl.dayfit.mossydevice.configuration
+package pl.dayfit.mossykeysync.configuration
 
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Profile
@@ -6,8 +6,8 @@ import org.springframework.web.socket.config.annotation.EnableWebSocket
 import org.springframework.web.socket.config.annotation.WebSocketConfigurer
 import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry
 import pl.dayfit.mossyauthstarter.configuration.properties.SecurityConfigurationProperties
-import pl.dayfit.mossydevice.ws.handler.AuthHandlerDecorator
-import pl.dayfit.mossydevice.ws.interceptor.QueryParamHandshakeInterceptor
+import pl.dayfit.mossykeysync.ws.handler.AuthHandlerDecorator
+import pl.dayfit.mossykeysync.ws.interceptor.QueryParamHandshakeInterceptor
 
 @Profile("!raw-websocket")
 @Configuration
