+++ mossy-key-sync/src/main/kotlin/pl/dayfit/mossykeysync/dto/response/InitKeySyncResponseDto.kt
@@ -1,4 +1,4 @@
-package pl.dayfit.mossydevice.dto.response
+package pl.dayfit.mossykeysync.dto.response
 
 data class InitKeySyncResponseDto(
     val code: String
