+++ mossy-key-sync/src/main/kotlin/pl/dayfit/mossykeysync/exception/RoleAlreadyInRoomException.kt
@@ -1,3 +1,3 @@
-package pl.dayfit.mossydevice.exception
+package pl.dayfit.mossykeysync.exception
 
 class RoleAlreadyInRoomException(override val message: String?) : RuntimeException()
