+++ mossy-key-sync/src/main/kotlin/pl/dayfit/mossykeysync/repository/redis/KeySyncRoomRepository.kt
@@ -1,8 +1,8 @@
-package pl.dayfit.mossydevice.repository.redis
+package pl.dayfit.mossykeysync.repository.redis
 
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
-import pl.dayfit.mossydevice.model.redis.KeySyncRoom
+import pl.dayfit.mossykeysync.model.redis.KeySyncRoom
 import java.util.UUID
 
 @Repository
