+++ mossy-key-sync/src/main/kotlin/pl/dayfit/mossykeysync/service/KeySyncService.kt
@@ -1,15 +1,15 @@
-package pl.dayfit.mossydevice.service
+package pl.dayfit.mossykeysync.service
 
 import org.springframework.stereotype.Service
 import org.springframework.web.socket.WebSocketSession
-import pl.dayfit.mossydevice.dto.response.InitKeySyncResponseDto
-import pl.dayfit.mossydevice.exception.RoleAlreadyInRoomException
-import pl.dayfit.mossydevice.model.redis.KeySyncRoom
-import pl.dayfit.mossydevice.repository.redis.KeySyncRoomRepository
-import pl.dayfit.mossydevice.type.KeySyncRole
-import pl.dayfit.mossydevice.ws.dto.WebSocketMessageDto
-import pl.dayfit.mossydevice.ws.dto.WebSocketServerMessageDto
-import pl.dayfit.mossydevice.ws.principal.DevicePrincipal
+import pl.dayfit.mossykeysync.dto.response.InitKeySyncResponseDto
+import pl.dayfit.mossykeysync.exception.RoleAlreadyInRoomException
+import pl.dayfit.mossykeysync.model.redis.KeySyncRoom
+import pl.dayfit.mossykeysync.repository.redis.KeySyncRoomRepository
+import pl.dayfit.mossykeysync.type.KeySyncRole
+import pl.dayfit.mossykeysync.ws.dto.WebSocketMessageDto
+import pl.dayfit.mossykeysync.ws.dto.WebSocketServerMessageDto
+import pl.dayfit.mossykeysync.ws.principal.DevicePrincipal
 import java.security.SecureRandom
 import java.util.UUID
