+++ mossy-key-sync/src/main/kotlin/pl/dayfit/mossykeysync/type/AuthFrameStatus.kt
@@ -1,4 +1,4 @@
-package pl.dayfit.mossydevice.type
+package pl.dayfit.mossykeysync.type
 
 enum class AuthFrameStatus {
     FAILED,
