+++ mossy-key-sync/src/main/kotlin/pl/dayfit/mossykeysync/type/KeySyncRole.kt
@@ -1,4 +1,4 @@
-package pl.dayfit.mossydevice.type
+package pl.dayfit.mossykeysync.type
 
 enum class KeySyncRole {
     RECEIVER, SENDER
