+++ mossy-key-sync/src/main/resources/application-debug.properties
@@ -1,2 +1,2 @@
-logging.level.pl.dayfit.mossydevice=DEBUG
+logging.level.pl.dayfit.mossykeysync=DEBUG
 logging.level.pl.dayfit.mossyauthstarter=DEBUG
